// client/src/types/api_response.ts
export type API = {
  ac: AircraftDetails[];
  ctime: number;
  total: number;
};

export type AircraftDetails = {
  hex: string;
  type: string;
  flight: string;
  r: string;
  t: string;
  dbFlags: number;
  alt_baro: number;
  alt_geom: number;
  gs: number;
  ias?: number;
  tas?: number;
  mach?: number;
  wd?: number;
  ws?: number;
  oat?: number;
  tat?: number;
  track: number;
  track_rate?: number;
  roll?: number;
  mag_heading?: number;
  true_heading?: number;
  baro_rate: number;
  geom_rate: number;
  squawk: string;
  emergency: string;
  category: string;
  nav_qnh?: number;
  nav_altitude_mcp?: number;
  nav_heading?: number;
  lat: number;
  lon: number;
  nic: number;
  rc: number;
  seen_pos: number;
  version: number;
  nic_baro: number;
  nac_p: number;
  nac_v: number;
  sil: number;
  sil_type: string;
  gva: number;
  sda: number;
  alert: number;
  spi: number;
  mlat: string[];
  tisb: string[];
  messages: number;
  seen: number;
  rssi: number;
};